FloatValue should be mapped to double type in MySQL (#337)

+++ core/src/main/java/com/scalar/db/storage/jdbc/JdbcDatabaseAdmin.java
@@ -49,7 +49,7 @@ public class JdbcDatabaseAdmin implements DistributedStorageAdmin {
                   .put(DataType.INT, "INT")
                   .put(DataType.BIGINT, "BIGINT")
                   .put(DataType.TEXT, "LONGTEXT")
-                  .put(DataType.FLOAT, "FLOAT")
+                  .put(DataType.FLOAT, "DOUBLE")
                   .put(DataType.DOUBLE, "DOUBLE")
                   .put(DataType.BOOLEAN, "BOOLEAN")
                   .put(DataType.BLOB, "LONGBLOB")
